HDF5Arrays for Julia

Overview

This repository ports (parts of) Bioconductor's HDF5Array package to provide a HDF5-backed array abstraction in Julia. Each HDF5Array instance only holds a pointer to the file in memory, and subsets of the data are retrieved on demand. This enables users to manipulate large datasets with minimal memory consumption. We support both dense arrays, represented as HDF5 datasets; and sparse matrices, represented using the 10X Genomics layout.

The HDF5Array class is implemented as a subtype of an AbstractArray and can (in theory) be used interchangeably with Arrays for any read operations. In practice, this abstraction is best used for mimicking the size and subsetting behavior of a real array. If values are needed, it is usually more efficient to manually extract large blocks of data from the file with extractdense() and extractsparse(), rather than relying on Base.getindex() to follow a sensible access pattern.

Write operations are not supported via the HDF5Array interface.

Quick start

Users may install this package from the GitHub repository through the usual process on the Pkg REPL:

add https://github.com/LTLA/HDF5Arrays.jl

And then:

julia> using HDF5Arrays

julia> temp = tempname();

julia> exampledense(temp, "foo", (50, 100));

julia> x = DenseHDF5Array(temp, "foo");

julia> size(x)
(50, 100)

julia> y = Array(x);

julia> typeof(y)
Matrix{Float64} (alias for Array{Float64, 2})

julia> sub = x[1:10,5:20];

julia> size(sub)
(10, 16)

And the same for sparse matrices:

julia> using HDF5Arrays, SparseArrays

julia> temp = tempname();

julia> examplesparse(temp, "foo", (50, 100), 0.1);

julia> x = SparseHDF5Matrix(temp, "foo");

julia> size(x)
(50, 100)

julia> y = SparseArrays.sparse(x);

julia> typeof(y)
SparseArrays.SparseMatrixCSC{Float64, Int64}

julia> sub = x[1:10,5:20];

julia> size(sub)
(10, 16)

The HDF5Array class

Class definition

HDF5Arrays.HDF5Array — Type

The HDF5Array is an abstract type that describes the concept of a HDF5-backed array abstraction. In other words, data is stored inside a HDF5 file and is retrieved on demand rather than being loaded into memory. T is the type of the data while N is the dimensionality.

Concrete subtypes include the DenseHDF5Array and the SparseHDF5Matrix. Subtypes are expected to implement the extractdense and extractsparse methods. They may also override the SparseArrays.issparse method to indicate whether they contain sparse data.

We provide conversion functions to quickly create in-memory Arrays or SparseMatrixCSC objects from any HDF5Arrays. In addition, subsetting operations will automatically views on the original array rather than immediately loading data from file. This enables lazy evaluation for memory-efficient operations on a subset of the dataset.

Currently, all HDF5Arrays are read-only; calling setindex! will fail.

Subsetting

Base.getindex — Method

getindex(x::HDF5Array{T,N}, I...)

Create a view into a HDF5Array{T,N} at the indices I. This does not read any data from file, only acting as a delayed subsetting operation.

This view-returning behavior continues when x is already a view on a HDF5Array. In that case, a new view is returned containing recomputed indices based on I and the parentindices(x).

Calling extractdense or extractsparse on this view will use the indices to extract the desired subset.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> y = x[1:5, :];

julia> isa(y, SubArray)
true

julia> z = y[1:2,1:5];

julia> isa(z, SubArray)
true

Extraction

HDF5Arrays.extractdense — Method

extractdense(x::SubArray{T,N,P<:HDF5Array{T,N},I,L}, indices...; blockdim = nothing)

Extract an in-memory dense Array from a SubArray x of a parent HDF5Array. The returned matrix contains the same values as x[indices...], in addition to any subsetting used to create the view on the parent.

blockdim is forwarded to the method for the parent HDF5Array.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> y = x[2:10, [2,3,5,9,10]];

julia> z = extractdense(y, :, :);

julia> size(z)
(9, 5)

HDF5Arrays.extractsparse — Method

extractsparse(x::SubArray{T,N,P<:HDF5Array{T,N},I,L}, i, j; blockdim = nothing)

Extract a sparse matrix from a SubArray x of a parent HDF5Array. The returned matrix contains the same values as x[i, j], in addition to any subsetting used to create the view on the parent.

blockdim is forwarded to the method for the parent HDF5Array.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.1)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> y = x[2:10, [2,3,5,9,10]];

julia> z = extractsparse(y, :, :);

julia> size(z)
(9, 5)

Conversions

Core.Array — Method

Array(x::HDF5Array{T,N}})

Convert a HDF5Array into an in-memory Array of the same data type and dimension. This is equivalent to using extractdense while requesting the full extent of each dimension.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> y = Array(x);

julia> size(y)
(20, 10)

Array(x::SubArray{T,N,P<:HDF5Array{T,N},I,L})

Convert a view of a HDF5Array into an in-memory Array of the same type and dimension. This is equivalent to using extractdense while requesting the full extent of each dimension.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> y = Array(x[1:10,2:5]);

julia> size(y)
(10, 4)

SparseArrays.sparse — Method

sparse(x::HDF5Array{T,N})

Convert a 2-dimensional HDF5Array into an in-memory sparse matrix of the same data type and dimension, assuming that T is some numeric or boolean type. This is equivalent to using extractsparse while requesting the full extent of each dimension. Note that this only makes sense if x contains a high proportion of zeros.

Examples

julia> using HDF5Arrays, SparseArrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.2)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> y = sparse(x);

julia> typeof(y)
SparseMatrixCSC{Float64, Int64}

SparseArrays.sparse — Method

sparse(x::SubArray{T,2,P<:HDF5Array{T,2},I,L})

Convert a view of a 2-dimensional HDF5Array into an in-memory sparse matrix of the same type and dimension. This is equivalent to using extractsparse while requesting the full extent of each dimension.

Examples

julia> using HDF5Arrays, SparseArrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.2)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> y = sparse(x[1:10,2:5]);

julia> typeof(y)
SparseMatrixCSC{Float64, Int64}

Other methods

Base.setindex! — Method

setindex!(x::HDF5Array{T,N}, v, I::Vararg{Integer,N})

All HDF5Arrays are strictly read-only, so any attempt to call this function will throw an error.

The DenseHDF5Array class

Class definition

HDF5Arrays.DenseHDF5Array — Type

This class implements a AbstractArray wrapper around a HDF5 dataset. Instances of this class can be used to represent a HDF5 dataset inside collections that expect an AbstractArray. Most importantly, this class does not load any data into memory on construction (and only minimal loading in its show() method). This allows users to manipulate arbitrarily large datasets in low-memory environments.

The DenseHDF5Array is structurally dense, reflecting the fact that a HDF5 dataset explicitly stores all values (zero or otherwise). T can be any type though this is usually numeric, and occasionally boolean. Any number of dimensions is supported, but note that the dimensions are permuted in Julia; the first dimension is the fastest changing in Julia but is the slowest changing in HDF5.

It is also possible to perform calculations on this class, in which case values are retrieved from file on demand. While memory-efficient, this approach is likely to be very slow, as it involves multiple (often redundant) I/O calls to the disk. We recommend using this class as a placeholder for a real array in code that does not require the actual values. Once values are explicitly needed, the entire array can be loaded into memory with extractdense or extractsparse.

Constructor

HDF5Arrays.DenseHDF5Array — Method

HDF5Array(file, name)

Create a HDF5Array for a dense dataset inside a HDF5 file file at name name. No values are loaded into memory by the constructor (though showing the matrix may load a few values for display).

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> size(x)
(20, 10)

julia> using SparseArrays

julia> SparseArrays.issparse(x)
false

Basic methods

Base.size — Method

size(x::DenseHDF5Array{T,N})

Get the size of a DenseHDF5Array{T,N} as a N-tuple.

size(x::SparseHDF5Matrix{Tv,Ti})

Get the size of a DenseHDF5Array{T,N} as a N-tuple.

Base.getindex — Method

getindex(x::DenseHDF5Array{T,N}, I::Vararg{Integer,N})

Get the value of a DenseHDF5Array{T,N} at the position specified by indices I. It would be unwise to use this function for anything other than show() - we suggest using extractdense to obtain larger blocks of data.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> typeof(getindex(x, 1, 1)) 
Float64

Extraction

HDF5Arrays.extractdense — Method

extractdense(x::DenseHDF5Array{T,N}, I...; blockdim = nothing)

Extract an in-memory dense Array from a DenseHDF5Array. The returned array contains the same values as x[I...].

For arbitrary indices, this function performs block-by-block extraction to reduce memory usage. The size of each block is determined by blockdim, which should be a tuple of length equal to the number of dimensions. If blockdim is not specified, the chunk dimensions are used instead; if the dataset is stored in a contiguous layout, arbitrary block dimensions are used.

Some optimization is applied if I only consists of AbstractRange{Int} values.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10))

julia> x = DenseHDF5Array(tmp, "stuff");

julia> y = extractdense(x, 1:5, 1:5);

julia> size(y)
(5, 5)

julia> y2 = extractdense(x, [1,3,5,7], [2,4,6,8,10]);

julia> size(y2)
(4, 5)

HDF5Arrays.extractsparse — Method

extractsparse(x::DenseHDF5Array{T,N}, i, j; blockdim = nothing)

Extract an in-memory sparse matrix from a 2-dimensional DenseHDF5Array. The returned matrix contains the same values as x[i, j]. This assumes that the type is either numeric or boolean, and is only useful when there is a high proportion of zero values in x.

For arbitrary indices, this function performs block-by-block extraction to reduce memory usage. The size of each block is determined by blockdim, which should be a tuple of length equal to the number of dimensions. If blockdim is not specified, the chunk dimensions are used instead; if the dataset is stored in a contiguous layout, arbitrary block dimensions are used.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (20, 10); density = 0.2)

julia> x = DenseHDF5Array(tmp, "stuff");

julia> y = extractsparse(x, 1:5, 1:5);

julia> typeof(y)
SparseArrays.SparseMatrixCSC{Float64, Int64}

The SparseHDF5Matrix class

Class definition

HDF5Arrays.SparseHDF5Matrix — Type

This class implements a AbstractArray wrapper around a sparse matrix in a HDF5 file. Instances of this class can be used to represent a HDF5 dataset inside collections that expect an AbstractArray. Most importantly, this class does not load any data into memory on construction (and only minimal loading in its show() method). This allows users to manipulate arbitrarily large datasets in low-memory environments.

HDF5 does not actually implement a native sparse format (at least, not of time of writing). Instead, we expect the 10X Genomics format where the matrix contents are stored in a group containing:

shape, an integer dataset of length 2 containing the number of rows and columns in the matrix.

data, a numeric dataset containing all of the non-zero elements in compressed sparse column (CSC) order.

indices, an integer dataset containing the 0-based row indices of all non-zero elements in CSC order.

indptr, an integer dataset containing the 0-based pointers into indices for the start and end of each column.

As with the DenseHDF5Array, we recommend using instances of this class as a placeholder for a real matrix, rather than computing directly on it. Once values are explicitly needed, the entire array can be loaded into memory with extractsparse.

Constructor

HDF5Arrays.SparseHDF5Matrix — Method

SparseHDF5Matrix(file, name)

Create a HDF5Array for a sparse matrix inside a HDF5 file file at group name. The group is expected to follow the 10X Genomics format, see SparseHDF5Matrix for more details.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.2)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> size(x)
(20, 10)

julia> using SparseArrays

julia> SparseArrays.issparse(x)
true

Basic methods

Base.size — Method

size(x::DenseHDF5Array{T,N})

Get the size of a DenseHDF5Array{T,N} as a N-tuple.

size(x::SparseHDF5Matrix{Tv,Ti})

Get the size of a DenseHDF5Array{T,N} as a N-tuple.

Base.getindex — Method

getindex(x::SparseHDF5Matrix{Tv,Ti}, i, j)

Get the value of x[i,j]. It would be unwise to use this function for anything other than show() - we suggest using extractsparse to obtain larger blocks of data.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.2)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> typeof(getindex(x, 1, 1))
Float64

SparseArrays.issparse — Method

issparse(x::SparseHDF5Matrix{Tv,Ti})

Returns true.

Extraction

HDF5Arrays.extractdense — Method

extractdense(x::SparseHDF5Matrix{Tv,Ti}, i, j; blockdim = nothing)

Extract an in-memory dense Matrix from a SparseHDF5Matrix. The returned matrix contains the same values as x[i, j].

blockdim is currently ignored and is only provided for consistency with the method for DenseHDF5Arrays.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.2)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> y = extractdense(x, 1:5, [2,3,6,7]);

julia> size(y)
(5, 4)

HDF5Arrays.extractsparse — Method

extractsparse(x::SparseHDF5Matrix{Tv,Ti}, i, j; blockdim = nothing)

Extract an in-memory sparse matrix from a SparseHDF5Matrix x. The returned matrix contains the same values as x[i, j].

blockdim is currently ignored and is only provided for consistency with the method for DenseHDF5Arrays.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (20, 10), 0.2)

julia> x = SparseHDF5Matrix(tmp, "stuff");

julia> y = extractsparse(x, 1:5, [2,3,6,7]);

julia> size(y)
(5, 4)

Miscellaneous

HDF5Arrays.exampledense — Method

exampledense(x, name, dims; density = 1, compress = 6, chunkdim = nothing)

Create an example dense dataset with name inside a HDF5 file/group. x may be either a string containing the name of a new HDF5 file, or a handle into an existing HDf5 file or group. dims should be a tuple specifying the dimensions of the dataset.

The HDF5 dataset is created with random double-precision values at the specified density. Lowering the density will introduce zero values randomly into the dataset. Note that this setting does not affect the structural density of the dataset.

compress specifies the DEFLATE compression level - if set to zero, no compression is used. chunkdim should be a tuple of length equal to dims, specifying the chunk dimensions to be used for compression. This is heuristically determined if not supplied, and is ignored if compress = 0.

Nothing is returned by this function.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> exampledense(tmp, "stuff", (10, 5, 2))

julia> y = DenseHDF5Array(tmp, "stuff");

julia> size(y)
(10, 5, 2)

HDF5Arrays.examplesparse — Method

examplesparse(x, name, dims, density; compress = 6, chunkdim = 10000)

Create an example sparse dataset with name inside a HDF5 file/group. x may be either a string containing the name of a new HDF5 file, or a handle into an existing HDf5 file or group. dims should be a tuple specifying the dimensions of the dataset.

The sparse matrix is created with random double-precision values at the specified density. Contents are written to file inside group name in the 10X Genomics format, which uses the compressed sparse column layout. Briefly: values are stored in data, row indices in indices, column pointers in indptr and the dimensions in shape.

compress specifies the DEFLATE compression level - if set to zero, no compression is used. chunkdim should be a tuple of length equal to dims, specifying the chunk dimensions to be used for compression. This is heuristically determined if not supplied, and is ignored if compress = 0.

Nothing is returned by this function.

Examples

julia> using HDF5Arrays

julia> tmp = tempname();

julia> examplesparse(tmp, "stuff", (10, 20), 0.1)

julia> y = SparseHDF5Matrix(tmp, "stuff");

julia> size(y)
(10, 20)

Contact

This package is maintained by Aaron Lun (@LTLA). If you have bug reports or feature requests, please post them as issues at the GitHub repository.